requirements.txt
joblib==1.2.0
mediapipe==0.10.3
numpy==1.24.3
opencv_contrib_python==4.8.0.76
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==9.5.0
Pillow==10.1.0
scikit_learn==1.2.2
streamlit==1.26.0
streamlit_lottie==0.0.5
